filter skill and type added to the search

+++ src/tech_test/search/views.py
@@ -2,29 +2,69 @@ from django.shortcuts import render, redirect
 from django.urls import reverse
 from django.http import HttpResponseRedirect
 from .models import Job
-from .API import search_job
+from .API import search_job, get_aggregators
 
 # Create your views here.
 def search(request):
-
+    Skill.objects.all().delete()
+    Type.objects.all().delete()
     context ={'job_search': '','place_search': '', 'remote_search': ''}
     return render(request, "search/search_home.html", context)
 
 def get_query(request):
 
+
+    context, skills_selected, types_selected = get_query_request_handle(request, get_aggregators())
+    jobs = search_job(size = 9, remote = True, skills = skills_selected, type = types_selected)
+    if len(jobs) != 0:
+
+        Job.objects.all().delete()
+        parse_jobs(jobs)
+        job_items = Job.objects.all()
+        
+        context['job_items'] = job_items
+
+    return render(request, "search/search_result.html",context)
+
+def get_query_request_handle(request, *args):
+    args = args[0]
+    context = {}
+    skills = []
+    types = []
     job_search = request.GET.get('job_search',"")
+    context['job_search'] = job_search
+
     place_search = request.GET.get('place_search',"")
+    context['place_search'] = place_search
+
     if request.GET.get('remote_search',"") == 'on':
         remote_search = 'checked'
     else:
         remote_search = ''
-    jobs = search_job(size = 9, remote = True, skills = ['Software Development'])
-    Job.objects.all().delete()
-    parse_jobs(jobs)
-    job_items = Job.objects.all()
-    context = {'job_search': job_search, 'place_search': place_search, 'remote_search': remote_search, 'job_items': job_items}
+    context['remote_search'] = remote_search
 
-    return render(request, "search/search_result.html",context)
+    for val in request.GET.keys():
+        if 'skill.' in val:
+            aux  = val.split(".", 1)[1]
+            skills.append(aux)
+        if 'type.' in val:
+            aux  = val.split(".", 1)[1]
+            types.append(aux)
+    
+
+
+    skills_checkbox = args[0]
+    for skill in skills_checkbox:
+        if skill.get('value') in skills:
+            skill['checked'] = 'checked'
+    context['skills_checkbox'] = skills_checkbox
+
+    type_checkbox = args[1] 
+    for type in type_checkbox:
+        if type.get('value') in types:
+            type['checked'] = 'checked'
+    context['type_checkbox'] = type_checkbox     
+    return context, skills, types
 
 def parse_jobs(jobs):
     for job in jobs:
@@ -43,23 +83,31 @@ def parse_jobs(jobs):
         for skill in job.get('skills'):
             skills.append(skill.get('name'))
         skills = ', '.join(skills)
-        if job.get('compensation').get('visible'):
-            min = str(job.get('compensation').get('data').get('minAmount'))
-            max = str(job.get('compensation').get('data').get('maxAmount'))
-            curr = str(job.get('compensation').get('data').get('currency'))
-            per = str(job.get('compensation').get('data').get('periodicity'))
-            if min == 'None':
-                max = max.split('.', 1)[0]
-                compensation = f'{max} {curr}/{per}'
-            elif max == 'None':
-                min = min.split('.', 1)[0]
-                compensation = f'{min} {curr}/{per}'
+        if job.get('compensation') != None:
+            if job.get('compensation').get('visible') != None:
+                try:
+                    min = str(job.get('compensation').get('data').get('minAmount'))
+                    max = str(job.get('compensation').get('data').get('maxAmount'))
+                    curr = str(job.get('compensation').get('data').get('currency'))
+                    per = str(job.get('compensation').get('data').get('periodicity'))
+                    if min == 'None':
+                        max = max.split('.', 1)[0]
+                        compensation = f'{max} {curr}/{per}'
+                    elif max == 'None':
+                        min = min.split('.', 1)[0]
+                        compensation = f'{min} {curr}/{per}'
+                    else:
+                        min = min.split('.', 1)[0]
+                        max = max.split('.', 1)[0]
+                        compensation = f'{min} - {max} {curr}/{per}'
+                except:
+                    pass
+                
             else:
-                min = min.split('.', 1)[0]
-                max = max.split('.', 1)[0]
-                compensation = f'{min} - {max} {curr}/{per}'
+                compensation = ''
         else:
             compensation = ''
         obj = Job.objects.create(_id = id, objective = objective, type = type, organization_name = organization_name,
         organization_picture = organization_picture, locations = locations, remote = remote, skills = skills,
         compensation = compensation)
+
